Updating for working export on mac os, working on widows, and easiest integration in eclipse
> I noticed that the 32bit version didn’t work on macos : i refactored
the native lib forlder to have a macos64 and a windows32 as inticated
in the processing wiki.
> I noticed that the macos export didn’t work due to the loading of
native lib : i made a dynamic loading of this lib for macos64 and
windows32 so the export work great now for both os, and the use of your
librarie is easier in eclipse. (my code i based on the same system of
the video librarie from Andrés Colubri)
> I noticed that the Cascad loading didn’t work on windows, so i
updated the dynamic loading, so now it works great and the use of your
librarie is easier in eclipse. (my code i based on the same system of
the video librarie from Andrés Colubri)